+++ rules.go
@@ -31,6 +31,7 @@ var defaultValidatorErrorMessages = map[string]string{
 	"or":        "ruleSets in $field did not pass based on or logic",
 	"xor":       "ruleSets in $field did not pass based on xor logic",
 	"choices":   "$value does not include in allowed choices: $choices",
+	"string":    "not a string value",
 
 	// Requires
 	"when_exist_one": "$field is required because at least one of $choices fields are not nil, empty or zero(0, \"\", '')",
@@ -264,4 +265,8 @@ func choicesRule(choices interface{}) func(interface{}) bool {
 
 		return false
 	}
+}
+
+func stringRule(input interface{}) bool {
+	return reflect.TypeOf(input).Kind() == reflect.String
 }